ZEN-34845: Release 6.8.0

+++ Products/ZenCallHome/transport/methods/versioncheck.py
@@ -7,30 +7,31 @@
 #
 ##############################################################################
 
-
 import json
 import logging
 import time
-from urllib import urlencode
+
 import urllib2
+
+from urllib import urlencode
 
 from Products.ZenUtils.Version import Version
 
-VERSION_CHECK_URL = 'https://callhome.zenoss.com/callhome/v2/versioncheck'
+VERSION_CHECK_URL = "https://callhome.zenoss.com/callhome/v2/versioncheck"
 _URL_TIMEOUT = 5
-logger = logging.getLogger('zen.callhome')
+logger = logging.getLogger("zen.callhome")
 
 
 def version_check(dmd):
-    params = urlencode({'product': dmd.getProductName()})
+    params = urlencode({"product": dmd.getProductName()})
     try:
         httpreq = urllib2.urlopen(VERSION_CHECK_URL, params, _URL_TIMEOUT)
         returnPayload = json.loads(httpreq.read())
     except Exception as e:
-        logger.warning('Error retrieving version from callhome server: %s', e)
+        logger.warning("Error retrieving version from callhome server: %s", e)
     else:
-        available = Version.parse('Zenoss ' + returnPayload['latest'])
+        available = Version.parse("Zenoss " + returnPayload["latest"])
         version = available.short()
-        dmd.lastVersionCheck = long(time.time())
-        if getattr(dmd, 'availableVersion', '') != version:
+        dmd.lastVersionCheck = int(time.time())
+        if getattr(dmd, "availableVersion", "") != version:
             dmd.availableVersion = version
